+++ src/helpers/getPlotHtmlAndWrite.ts
@@ -1,0 +1,25 @@
+import { JSDOM } from "jsdom"
+import fs from "fs"
+import { Options } from "../types/SimpleData.types.js"
+import log from "./log.js"
+import { plot } from "@observablehq/plot"
+
+export default function getPlotHtmlAndWrite(plotOptions: any, path: string, options: Options) {
+
+    options.logs && log(plotOptions)
+
+    if (global.document === undefined) {
+        const jsdom = new JSDOM("")
+        global.document = jsdom.window.document
+    }
+
+    const chart = plot(plotOptions).outerHTML
+
+    fs.writeFileSync(
+        path,
+        chart.replace("<svg", '<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink"')
+    )
+
+    options.logs && log(`=> chart save to ${path}`, "blue")
+
+}
