+++ src/methods/saveChart.ts
@@ -1,21 +1,15 @@
 import { Options, SimpleDataItem } from "../types/SimpleData.types.js"
-import * as Plot from "@observablehq/plot"
-import { JSDOM } from "jsdom"
-import fs from "fs"
-import log from "../helpers/log.js";
+import { dot, line, barY, boxY, plot } from "@observablehq/plot"
+import getExtension from "../helpers/getExtension.js";
+import getPlotHtmlAndWrite from "../helpers/getPlotHtmlAndWrite.js";
 
 export default function saveChart(data: SimpleDataItem[], path: string, type: "dot" | "line" | "bar" | "box", x: any[], y: any[], color: string, options: Options): SimpleDataItem[] {
 
-    const extensionSplit = path.split(".")
-    const extension = extensionSplit[extensionSplit.length - 1].toLocaleLowerCase()
+    const extension = getExtension(path)
     if (extension !== "html") {
         throw new Error("For the moment, you can only export charts with file extension .html")
     }
 
-    if (global.document === undefined) {
-        const jsdom = new JSDOM("")
-        global.document = jsdom.window.document
-    }
     const markOption: { [key: string]: any } = { x, y }
 
     if (color && ["dot", "bar", "box"].includes(type)) {
@@ -26,13 +20,13 @@ export default function saveChart(data: SimpleDataItem[], path: string, type: "d
 
     let mark
     if (type === "dot") {
-        mark = Plot.dot(data, markOption)
+        mark = dot(data, markOption)
     } else if (type === "line") {
-        mark = Plot.line(data, markOption)
+        mark = line(data, markOption)
     } else if (type === "bar") {
-        mark = Plot.barY(data, markOption)
+        mark = barY(data, markOption)
     } else if (type === "box") {
-        mark = Plot.boxY(data, markOption)
+        mark = boxY(data, markOption)
     } else {
         throw new Error("Unknown chart type.")
     }
@@ -48,14 +42,7 @@ export default function saveChart(data: SimpleDataItem[], path: string, type: "d
         }
     }
 
-    options.logs && log(plotOptions)
-
-    const plot = Plot.plot(plotOptions).outerHTML
-
-    fs.writeFileSync(
-        path,
-        plot.replace("<svg", '<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink"')
-    )
+    getPlotHtmlAndWrite(plotOptions, path, options)
 
     return data
 }
